Code cleanup and refactoring.
